Adding C source for network interface details
The bare bones Lua standard library does not have anything to identify
network or ethernet interface details. To not rely on external
libraries for such a primitive taks, it's better to just write some C
bindings to get the job done. Warning: the C-Lua binding part is yet
to be done as of now.